+++ src/app/set.ts
@@ -14,7 +14,7 @@ export async function setApp(appId: string, options: Options) {
   appId = getAppId(appId, extConfig?.config)
 
   if (!options.apikey) {
-    log.error(`Missing API key, you need to provide a API key to upload your bundle`)
+    log.error(`Missing API key, you need to provide an API key to upload your bundle`)
     program.error(``)
   }
   if (!appId) {