+++ src/bundle/compatibility.ts
@@ -28,7 +28,7 @@ export async function checkCompatibilityCommand(appId: string, options: Options)
     }
 
     if (!options.apikey) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     if (!appId) {