+++ src/bundle/delete.ts
@@ -18,7 +18,7 @@ export async function deleteBundle(bundleId: string, appId: string, options: Opt
     appId = getAppId(appId, extConfig?.config)
 
     if (!options.apikey) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     if (!appId) {