+++ src/bundle/unlink.ts
@@ -38,7 +38,7 @@ export async function unlinkDevice(channel: string, appId: string, options: Opti
     bundle = bundle || packVersion
 
     if (!options.apikey) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     if (!appId) {