+++ src/channel/add.ts
@@ -19,7 +19,7 @@ export async function addChannel(channelId: string, appId: string, options: Opti
     appId = getAppId(appId, extConfig?.config)
 
     if (!options.apikey) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     if (!appId) {