+++ src/channel/delete.ts
@@ -20,7 +20,7 @@ export async function deleteChannel(channelId: string, appId: string, options: D
     appId = getAppId(appId, extConfig?.config)
 
     if (!options.apikey) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     if (!appId) {