+++ src/channel/list.ts
@@ -14,7 +14,7 @@ export async function listChannels(appId: string, options: OptionsBase) {
     appId = getAppId(appId, extConfig?.config)
 
     if (!options.apikey)
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
 
     if (!appId) {
       log.error('Missing argument, you need to provide a appId, or be in a capacitor project')