+++ src/channel/set.ts
@@ -50,7 +50,7 @@ export async function setChannel(channel: string, appId: string, options: Option
     appId = getAppId(appId, extConfig?.config)
 
     if (!options.apikey) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     if (!appId) {