+++ src/login.ts
@@ -22,7 +22,7 @@ export async function login(apikey: string, options: Options, shouldExit = true)
 
   if (!apikey) {
     if (shouldExit) {
-      log.error('Missing API key, you need to provide a API key to upload your bundle')
+      log.error('Missing API key, you need to provide an API key to upload your bundle')
       program.error('')
     }
     return false